+++ src/components/LoanCollections/KivaClassicSingleCategoryCarousel.vue
@@ -1,10 +1,10 @@
 <template>
 	<div>
 		<div>
-			<h2 class="tw-mb-2 tw-inline">
+			<h2 class="tw-mb-2 tw-block md:tw-inline">
 				{{ name }}
 			</h2>
-			<kv-pill class="tw-ml-4.5" v-if="climateChallenge">
+			<kv-pill class="md:tw-ml-4.5" v-if="climateChallenge">
 				<template #icon>
 					<icon-climate-challenge class="tw-h-3 tw-w-2 tw-mr-0.5" />
 				</template>
@@ -33,6 +33,7 @@
 import IconClimateChallenge from '@/assets/icons/inline/eco-challenge/globe-leaf.svg';
 import KivaClassicLoanCarousel from '@/components/LoanCollections/KivaClassicLoanCarousel';
 import KvPill from '@/components/Kv/KvPill';
+import { FLSS_ORIGIN_NOT_SPECIFIED } from '@/util/flssUtils';
 
 import {
 	getLoanChannel,
@@ -98,6 +99,10 @@ export default {
 			type: Boolean,
 			default: false
 		},
+		queryContext: {
+			type: String,
+			default: FLSS_ORIGIN_NOT_SPECIFIED,
+		}
 	},
 	data() {
 		return {
@@ -155,12 +160,13 @@ export default {
 				offset: 0,
 				limit: this.loanDisplaySettings?.loanLimit ?? 1,
 				basketId: this.cookieStore.get('kvbskt'),
+				origin: this.queryContext
 			};
 			const channelData = await getLoanChannel(
 				this.apollo,
 				this.loanChannelQueryMap,
 				channelUrl,
-				loanQueryVars
+				loanQueryVars,
 			);
 			const loanChannelData = channelData?.data?.lend?.loanChannelsById ?? [];
 			this.selectedChannel = loanChannelData?.[0];