+++ src/pages/Lend/LendByCategoryPage.vue
@@ -6,9 +6,9 @@
 
 		<!-- Eco Challenge CTA -->
 		<div v-if="ecoChallengeExp" class="tw-bg-brand-50">
-			<div class="tw-max-w-5xl tw-mx-auto lg:tw-px-6 tw-py-2.5 lg:tw-py-4">
-				<div class="tw-flex">
-					<eco-challenge-badge class="tw-h-14 tw-w-12 tw-mr-4 tw-flex-none" />
+			<div class="tw-max-w-5xl tw-mx-auto lg:tw-px-6 tw-p-2.5 lg:tw-py-4">
+				<div class="md:tw-flex">
+					<eco-challenge-badge class="tw-h-14 tw-w-12 md:tw-mr-4 tw-flex-none tw-mx-auto" />
 					<div class="tw-w-full">
 						<h2>Join Kiva’s October eco-friendly challenge!</h2>
 						<p class="tw-text-subhead tw-mt-2">
@@ -124,6 +124,7 @@ import { addCustomChannelInfo } from '@/util/categoryUtils';
 import logReadQueryError from '@/util/logReadQueryError';
 import { readJSONSetting } from '@/util/settingsUtils';
 import { indexIn } from '@/util/comparators';
+import { FLSS_ORIGIN_LEND_BY_CATEGORY } from '@/util/flssUtils';
 import { isLoanFundraising } from '@/util/loanUtils';
 import experimentQuery from '@/graphql/query/experimentAssignment.graphql';
 import basketItems from '@/graphql/query/basketItems.graphql';
@@ -399,6 +400,7 @@ export default {
 					excludeIds: ssrLoanIds,
 					imgDefaultSize: this.showHoverLoanCards ? 'w480h300' : 'w480h360',
 					imgRetinaSize: this.showHoverLoanCards ? 'w960h600' : 'w960h720',
+					origin: FLSS_ORIGIN_LEND_BY_CATEGORY
 					// @todo variables for fetching data for custom channels
 				},
 			}).then(({ data }) => {
@@ -417,6 +419,7 @@ export default {
 						ids: this.realCategoryIds,
 						imgDefaultSize: this.showHoverLoanCards ? 'w480h300' : 'w480h360',
 						imgRetinaSize: this.showHoverLoanCards ? 'w960h600' : 'w960h720',
+						origin: FLSS_ORIGIN_LEND_BY_CATEGORY
 					},
 				}).subscribe({
 					next: ({ data }) => {
@@ -499,6 +502,7 @@ export default {
 					ids: recLoanChannels.map(channel => channel.id),
 					imgDefaultSize: this.showHoverLoanCards ? 'w480h300' : 'w480h360',
 					imgRetinaSize: this.showHoverLoanCards ? 'w960h600' : 'w960h720',
+					origin: FLSS_ORIGIN_LEND_BY_CATEGORY
 				};
 
 				return this.apollo.query({
@@ -637,7 +641,8 @@ export default {
 							ids: [category.id],
 							imgDefaultSize: this.showHoverLoanCards ? 'w480h300' : 'w480h360',
 							imgRetinaSize: this.showHoverLoanCards ? 'w960h600' : 'w960h720',
-							excludeIds
+							excludeIds,
+							origin: FLSS_ORIGIN_LEND_BY_CATEGORY
 						};
 						return this.apollo.query({
 							query: recommendedLoansQuery,
@@ -678,7 +683,8 @@ export default {
 								ids: [category.id],
 								imgDefaultSize: this.showHoverLoanCards ? 'w480h300' : 'w480h360',
 								imgRetinaSize: this.showHoverLoanCards ? 'w960h600' : 'w960h720',
-								excludeIds
+								excludeIds,
+								origin: FLSS_ORIGIN_LEND_BY_CATEGORY
 							},
 						}).then(({ data }) => {
 							const fetchedCategory = this.categoryServiceExpActive