+++ src/util/loanSearch/searchStateUtils.js
@@ -29,6 +29,8 @@ export function getValidatedSearchState(loanSearchState, allFacets, queryType) {
 
 	const validatedThemeIds = loanSearchState?.themeId?.filter(t => allFacets.themeIds.includes(t)) ?? [];
 
+	const validatedTagIds = loanSearchState?.tagId?.filter(t => allFacets.tagIds.includes(t)) ?? [];
+
 	const validatedPageOffset = isNumber(loanSearchState?.pageOffset)
 		? loanSearchState.pageOffset
 		: defaultLoanSearchState.pageOffset;
@@ -37,14 +39,21 @@ export function getValidatedSearchState(loanSearchState, allFacets, queryType) {
 		? loanSearchState.pageLimit
 		: defaultLoanSearchState.pageLimit;
 
+	const validatedDistributionModel = allFacets.distributionModels
+		.includes(loanSearchState?.distributionModel?.toUpperCase())
+		? loanSearchState.distributionModel
+		: null;
+
 	return {
 		gender: validatedGender,
 		countryIsoCode: validatedIsoCodes,
 		sectorId: validatedSectorIds,
 		sortBy: validatedSortBy,
 		themeId: validatedThemeIds,
+		tagId: validatedTagIds,
 		pageOffset: validatedPageOffset,
 		pageLimit: validatedPageLimit,
+		distributionModel: validatedDistributionModel,
 	};
 }